+++ src/wallrotate/resources/check_fullscreen.js
@@ -1,0 +1,3 @@
+var w = workspace.activeWindow;
+var fs = w ? w.fullScreen : false;
+callDBus("com.wallrotate.FullscreenCheck", "/FullscreenCheck", "", "Report", fs);